+++ plugins/woocommerce-admin/client/launch-your-store/hub/main-content/pages/loading.tsx
@@ -1,29 +1,12 @@
 /**
  * External dependencies
  */
-import clsx from 'clsx';
-/**
- * Internal dependencies
- */
-import type { MainContentComponentProps } from '../xstate';
-export const LoadingPage = ( props: MainContentComponentProps ) => {
+import { Spinner } from '@woocommerce/components';
+
+export const LoadingPage = () => {
 	return (
-		<div
-			className={ clsx(
-				'launch-store-loading-page__container',
-				props.className
-			) }
-		>
-			<p>Main Content - Loading page</p>
-			<button
-				onClick={ () => {
-					props.sendEventToSidebar( {
-						type: 'LAUNCH_STORE_SUCCESS',
-					} );
-				} }
-			>
-				Launch Store Success
-			</button>
+		<div className="spinner-container">
+			<Spinner></Spinner>
 		</div>
 	);
 };
