+++ plugins/woocommerce/changelog/fix-lys-refetch-site-cache-status
@@ -1,0 +1,4 @@
+Significance: patch
+Type: fix
+
+Fetch site cache status correctly if directly navigating to LYS Success page, and some refactoring
